+++ example/db/schema.rb
@@ -37,6 +37,7 @@ ActiveRecord::Schema.define(version: 2018_11_12_080414) do
     t.string "invitable_type", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
+    t.index ["slug"], name: "index_users_on_slug", unique: true
   end
 
 end