+++ spec/checkers/belongs_to_presence_checker_spec.rb
@@ -43,7 +43,7 @@ RSpec.describe DatabaseConsistency::Checkers::BelongsToPresenceChecker do
     end
   end
 
-  context 'when foreign key is missed' do
+  context 'when foreign key is missing' do
     before do
       define_database do
         create_table :countries